added hybrid mode of Xeon and Fermi for double precision
git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/trunk@1029 4eea3252-f0fb-4393-894d-40516dce545b

+++ src/sgpp/basis/linear/noboundary/operation/datadriven/OperationBIterativeSPHybridSSEOCLLinear.cpp
@@ -203,7 +203,7 @@ double OperationBIterativeSPHybridSSEOCLLinear::multVectorized(DataVectorSP& alp
 								float index_calc = eval - (ptrIndex[(j*dims)+d]);
 								float abs = fabs(index_calc);
 								float last = 1.0f - abs;
-								float localSupport = std::max<double>(last, 0.0f);
+								float localSupport = std::max<float>(last, 0.0f);
 								curSupport *= localSupport;
 							}
